requirements.txt
pandas==2.2.3
streamlit==1.40.2
streamlit-option-menu==0.4.0
pillow==11.0.0
babel==2.16.0
plotly==5.24.1